Lint, vendorize
