--- _config.php
@@ -1,2 +1,0 @@
-<?php
-define('LINK_FIELD_DIR', basename(__DIR__));
