+++ phpcs.xml.dist
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<ruleset name="PSR2">
+    <rule ref="PSR2">
+        <!-- Exclude camel caps methods - e.g. Controller::Form() -->
+        <exclude name="PSR1.Methods.CamelCapsMethodName" />
+        <exclude name="PSR1.Classes.ClassDeclaration" />
+		<exclude name="Squiz.Commenting.FunctionComment" />
+     </rule>
+
+     <rule ref="Generic.Files.LineLength">
+        <properties>
+            <property name="lineLimit" value="120"/>
+            <property name="absoluteLineLimit" value="120"/>
+        </properties>
+    </rule>
+
+    <rule ref="Squiz.Commenting.FunctionComment" />
+</ruleset>
